feat: Completed the mvp of this event

+++ lib/app/data/models/avaluation.dart
@@ -1,6 +1,6 @@
 class Avaluation {
   final String team;
-  final int grade;
+  final double grade;
 
   Avaluation({required this.team, required this.grade});
